try to be smarter about which pairs are tried
This mostly applies to the case when a station is run through the whole
set of currently fixed stations and isn't yet matched (so it goes back
on the queue). When it comes off again, we can skip the ones we tried
the previous time.

Has little (no?) effect on the sample.